User Guide

Structs

Main.XIMC.DeviceId — Type

DeviceId = device_t

Stores the id of a motor as int

Main.XIMC.Position — Type

Position = get_position_t

Stores the position of a motor.

Fields

Position::Cint: The step of the motor

uPosition::Cint The micro step of the motor

EncPosition::long_t The encoder position

Functions

Control

Functions to control the motors.

Main.XIMC.requestDevices — Function

requestDevices(address, stagenames...)

Open the connection to motors with given identifiers stagenames. If successful, returns a Vector{DeviceId} of given motors in the same order.

Main.XIMC.closeDevice — Function

closeDevice(device::DeviceId)

Close the connection to the motor.

closeDevice(device::DeviceId)

Close the connection to multiple motors.

Main.XIMC.closeDevices — Function

const closeDevices(devices::Vector{DeviceId}) = closeDevice.(devices)

Main.XIMC.getPosition — Function

getPosition(device::DeviceId)

Return the current position of specified motor in steps.

getPosition(devices::Vector{DeviceId}; fmt::Type=Matrix)

Return the current position of multiple motors.

The return type can be given by fmt:

Matrix: (#devices, 3)-Matrix ,where each row corresponds to the motor in devices and contains [Position, uPosition, EncPosition]

Vector: Vector of Tuple{Int,Int} of length #devices, where the tuple is (Position, uPosition)

The default is Matrix.

getPosition(devices::DeviceId...; fmt::Type=Matrix)

function getPosition(devices::Vector{DeviceId},
    cal::Dict{String,Tuple{Symbol,Float64}};
    outputunit=:m)
    
Return the position in given unit instead of steps.

Main.XIMC.getPosition! — Function

getPosition!(pos::Position, device::DeviceId)

Store the current position of specified motor into pos.

Main.XIMC.getSpeed — Function

getSpeed(device::DeviceId)

Returns a tuple (Speed, uSpeed) of the speed in steps/s.

Main.XIMC.setSpeed — Function

setSpeed(device::DeviceId,speed::Integer; info=false)

Set the speed of the motor in steps/s.

setSpeed(device::DeviceId,speed::Integer, uSpeed::Integer; info=false)

Set the speed of the motor in steps/s and (micro steps)/s.

setSpeed(device::Vector{DeviceId}, speed::Integer; info=false)

Set the speed of multiple motors in steps/s.

setSpeed(device::Vector{DeviceId}, speed::Integer; info=false)

Set the speed of multiple motors in steps/s and (micro steps)/s.

Main.XIMC.commandMove — Function

commandMove(device::DeviceId, pos::Real, upos::Real; info=false)

Move the motor to the Position given by the steps pos and micro steps upos. If info=true the return is printed to the console.

commandMove(device::DeviceId,pos::Position; info=false)

commandMove(devices::Vector{DeviceId},positions::Vector{Tuple{Int,Int}})

Move multiple motors to the given positions.

commandMove(devices::Vector{DeviceId},positions::Matrix{Int32})

commandMove(device::DeviceId,x::Real,cal::Dict{String,Tuple{Symbol,Float64}}; info=false,inputunit=:mm)

Move the motor using specified unit instead of steps.

commandMove(devices::Vector{DeviceId},x::Vector{<:Real},cal::Dict{String,Tuple{Symbol,Float64}}; info=false,inputunit=:mm)

Move multiplied motors using specified unit instead of steps.

Main.XIMC.commandWaitForStop — Function

commandWaitForStop(device::DeviceId; interval::UInt32=0x00000a,info::Bool=false)

Waiting for the specified motor to stop moving.

interval specifies how many milliseconds the function waits to check again whether the motor is still moving.

commandWaitForStop(devices::Vector{DeviceId}; interval::UInt32=0x00000a,info::Bool=false)

Waiting for multiple motors to stop moving.

Information

Functions which print out information to the console.

Main.XIMC.infoXIMC — Function

infoXIMC()

Prints and returns the current version number of the XIMC library.

Main.XIMC.infoDevice — Function

infoDevice(device::DeviceId)

Print basic information about the motor to the console.

Main.XIMC.infoMoveSettings — Function

infoMoveSettings(device::DeviceId)

Print move settings of the motor to the console.

Main.XIMC.infoPosition — Function

infoPosition(device::DeviceId)

Print position of the motor to the console.

Main.XIMC.infoSerial — Function

infoSerial(device::DeviceId)

Print serial number of the motor to the console.

Main.XIMC.infoSpeed — Function

infoSpeed(device::DeviceId)

Print speed settings of the motor to the console.

Main.XIMC.infoStage — Function

infoStage(device::DeviceId)

Print stage information about the motor to the console.

Main.XIMC.infoStatus — Function

infoStatus(device::DeviceId)

Print status information about the motor to the console.

Miscellaneous

Main.XIMC.isSamePosition — Function

isSamePosition(x::Position, y::Position)

Returns wether x and y are the same Position.

Main.XIMC.isGreaterEqPosition — Function

isGreaterEqPosition(x::Position, y::Position)

Returns x >= y. Note, that you can use the operator itself (eg. x >= y).

Main.XIMC.isSmallerEqPosition — Function

isSmallerEqPosition(x::Position, y::Position)

Returns x <= y. Note, that you can use the operator itself (eg. x <= y).

Main.XIMC.isGreaterPosition — Function

isGreaterPosition(x::Position, y::Position)

Returns x > y. Note, that you can use the operator itself (eg. x > y).

Main.XIMC.isSmallerPosition — Function

isSmallerPosition(x::Position, y::Position)

Returns x < y. Note, that you can use the operator itself (eg. x < y).

Base.:== — Function

const ==(x::Position, y::Position) = isSamePosition(x, y)

Base.:>= — Function

const >=(x::Position, y::Position) = isGreaterEqPosition(x, y)

Base.:<= — Function

const <=(x::Position, y::Position) = isSmallerEqPosition(x, y)

Base.:> — Function

const >(x::Position, y::Position) = isGreaterPosition(x, y)

Base.:< — Function

const <(x::Position, y::Position) = isSmallerPosition(x, y)

Index

Main.XIMC.DeviceId

Main.XIMC.Position

Base.:<

Base.:<=

Base.:==

Base.:>

Base.:>=

Main.XIMC.closeDevice

Main.XIMC.closeDevices

Main.XIMC.commandMove

Main.XIMC.commandWaitForStop

Main.XIMC.getPosition

Main.XIMC.getPosition!

Main.XIMC.getSpeed

Main.XIMC.infoDevice

Main.XIMC.infoMoveSettings

Main.XIMC.infoPosition

Main.XIMC.infoSerial

Main.XIMC.infoSpeed

Main.XIMC.infoStage

Main.XIMC.infoStatus

Main.XIMC.infoXIMC

Main.XIMC.isGreaterEqPosition

Main.XIMC.isGreaterPosition

Main.XIMC.isSamePosition

Main.XIMC.isSmallerEqPosition

Main.XIMC.isSmallerPosition

Main.XIMC.requestDevices

Main.XIMC.setSpeed